Combine auth principal store to provider
