--- pkg/auth/dependency/principal/principal.go
@@ -1,15 +1,0 @@
-package principal
-
-import "github.com/skygeario/skygear-server/pkg/server/uuid"
-
-type Principal struct {
-	ID       string
-	Provider string
-	UserID   string
-}
-
-func New() Principal {
-	return Principal{
-		ID: uuid.New(),
-	}
-}
