+++ pkg/auth/dependency/provider/password/principal.go
@@ -1,0 +1,16 @@
+package password
+
+import "github.com/skygeario/skygear-server/pkg/server/uuid"
+
+type Principal struct {
+	ID            string
+	UserID        string
+	AuthData      interface{}
+	PlainPassword string
+}
+
+func NewPrincipal() Principal {
+	return Principal{
+		ID: uuid.New(),
+	}
+}
