+++ pkg/auth/handler/signup.go
@@ -7,8 +7,6 @@ import (
 
 	"github.com/skygeario/skygear-server/pkg/auth/dependency/provider/password"
 
-	"github.com/skygeario/skygear-server/pkg/auth/dependency/principal"
-
 	"github.com/skygeario/skygear-server/pkg/auth"
 	"github.com/skygeario/skygear-server/pkg/auth/dependency"
 	"github.com/skygeario/skygear-server/pkg/auth/response"
@@ -72,7 +70,6 @@ type SignupHandler struct {
 	PasswordChecker      dependency.PasswordChecker `dependency:"PasswordChecker"`
 	TokenStore           authtoken.Store            `dependency:"TokenStore"`
 	AuthInfoStore        authinfo.Store             `dependency:"AuthInfoStore"`
-	AuthPrincipalStore   principal.Store            `dependency:"AuthPrincipalStore"`
 	PasswordAuthProvider password.Provider          `dependency:"PasswordAuthProvider"`
 }
 
@@ -123,44 +120,25 @@ func (h SignupHandler) Handle(req interface{}, _ handler.AuthContext) (resp inte
 		return
 	}
 
-	principal := principal.New()
+	// Create Principal
+	principal := password.NewPrincipal()
 
 	if payload.isAnonymous() {
 		panic("Unsupported signup anonymously")
-		// info = authinfo.NewAnonymousAuthInfo()
 	} else if payload.Provider != "" {
 		panic("Unsupported signup with provider")
-		// 	// Get AuthProvider and authenticates the user
-		// 	logger.Debugf(`Client requested auth provider: "%v".`, p.Provider)
-		// 	authProvider, err := h.ProviderRegistry.GetAuthProvider(p.Provider)
-		// 	if err != nil {
-		// 		response.Err = skyerr.NewInvalidArgument(err.Error(), []string{"provider"})
-		// 		return
-		// 	}
-		// 	principalID, providerAuthData, err := authProvider.Login(payload.Context(), p.ProviderAuthData)
-		// 	if err != nil {
-		// 		response.Err = skyerr.NewError(skyerr.InvalidCredentials, "unable to login with the given credentials")
-		// 		return
-		// 	}
-		// 	logger.Infof(`Client authenticated as principal: "%v" (provider: "%v").`, principalID, p.Provider)
-
-		// 	// Create new user info and set updated auth data
-		// 	info = skydb.NewProviderInfoAuthInfo(principalID, providerAuthData)
 	} else {
-		principal.Provider = "password"
 		principal.UserID = info.ID
+		principal.AuthData = payload.AuthData
+		principal.PlainPassword = payload.Password
 	}
 
-	err = h.AuthPrincipalStore.CreatePrincipal(principal)
+	err = h.PasswordAuthProvider.CreatePrincipal(principal)
 	if err != nil {
 		return
 	}
 
-	err = h.createProviderEntry(payload, principal)
-	if err != nil {
-		return
-	}
-
+	// Create auth token
 	tkn, err := h.TokenStore.NewToken(authContext.AuthInfo.ID)
 	if err != nil {
 		panic(err)
@@ -183,20 +161,6 @@ func (h SignupHandler) Handle(req interface{}, _ handler.AuthContext) (resp inte
 
 	// TODO: Update user last login time
 	// TODO: Audit
-
-	return
-}
-
-func (h SignupHandler) createProviderEntry(payload SignupRequestPayload, prin principal.Principal) (err error) {
-	if payload.isAnonymous() {
-		return
-	}
-
-	if payload.Provider != "" {
-		panic("Unsupported signup with provider")
-	} else {
-		err = h.PasswordAuthProvider.CreateEntry(prin.ID, payload.AuthData, payload.Password)
-	}
 
 	return
 }
